backend/requirements.txt
django==3.2
djangorestframework==3.12.4
environs==9.3.1
dj-database-url==0.5.0